+++ app/pages/login/login_page.php
@@ -7,38 +7,7 @@
 		<link href="<?php echo pf_config::get('base_url'); ?>app/assets/css/main.min.css" rel="stylesheet" media="screen">
                 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 		<link href="<?php echo pf_config::get('base_url'); ?>app/assets/css/main-responsive.min.css" rel="stylesheet">
-                
-                <style>
-			body {
-				margin-top: 20px;
-				background-color: #f5f5f5;
-			}
-			
-			.form-signin {
-        		max-width: 300px;
-       			padding: 19px 29px 29px;
-        		margin: 0 auto 20px;
-        		background-color: #fff;
-        		border: 1px solid #e5e5e5;
-        		-webkit-border-radius: 5px;
-           		-moz-border-radius: 5px;
-                border-radius: 5px;
-        		-webkit-box-shadow: 0 1px 2px rgba(0,0,0,.05);
-           		-moz-box-shadow: 0 1px 2px rgba(0,0,0,.05);
-                box-shadow: 0 1px 2px rgba(0,0,0,.05);
-      		}
-      		
-      		.form-signin .form-signin-heading, .form-signin .checkbox {
-       		 	margin-bottom: 10px;
-      		}
-      		
-      		.form-signin input[type="text"], .form-signin input[type="password"] {
-       			font-size: 16px;
-        		height: auto;
-        		margin-bottom: 15px;
-        		padding: 7px 9px;
-      		}
-		</style>
+                <link href="<?php echo pf_config::get('base_url'); ?>app/assets/css/style.css" rel="stylesheet">
 	</head>
         <body>
             <div class="container-fluid"> <!-- START CONTAINER -->
@@ -50,6 +19,8 @@
                                     
             <div id="login" class="" style="padding-top:100px;">
                 <form id="login" action="<?php echo pf_config::get('base_url').pf_config::get('index_page')?>/login/action" method="POST">
+                        
+                    <legend>Please Login</legend>
                     
                     <label for="username">
                         Username:
@@ -60,6 +31,7 @@
                         <input type="password" name="password" placeholder="Password" required />
                     </label>
                     <input class="button" type="submit" value="Login"/>
+                    
                 </form>
             </div>